+++ app/test/e2e/microservice-delete.spec.js
@@ -3,10 +3,11 @@ const chai = require('chai');
 
 const MicroserviceModel = require('models/microservice.model');
 const EndpointModel = require('models/endpoint.model');
-const UserModel = require('plugins/sd-ct-oauth-plugin/models/user.model');
 
-const { createUserAndToken, createMicroservice, createEndpoint } = require('./utils/helpers');
-const { getTestAgent, closeTestAgent } = require('./test-server');
+const {
+    createUserAndToken, createMicroservice, createEndpoint, mockGetUserFromToken
+} = require('./utils/helpers');
+const { getTestAgent, closeTestAgent } = require('./utils/test-server');
 
 chai.should();
 
@@ -23,13 +24,14 @@ describe('Delete a microservice', () => {
     });
 
     beforeEach(async () => {
-        await UserModel.deleteMany({}).exec();
         await MicroserviceModel.deleteMany({}).exec();
         await EndpointModel.deleteMany({}).exec();
     });
 
     it('Deleting a microservice should delete endpoints and the microservice document from the database (happy case)', async () => {
-        const { token } = await createUserAndToken({ role: 'ADMIN' });
+        const { token, user } = await createUserAndToken({ role: 'ADMIN' });
+
+        mockGetUserFromToken(user, token);
 
         const testMicroserviceOne = {
             name: `test-microservice-one`,
@@ -117,7 +119,9 @@ describe('Delete a microservice', () => {
     });
 
     it('Deleting a microservice should delete exclusive endpoints and the microservice document from the database, but preserve shared endpoints without redirects', async () => {
-        const { token } = await createUserAndToken({ role: 'ADMIN' });
+        const { token, user } = await createUserAndToken({ role: 'ADMIN' });
+
+        mockGetUserFromToken(user, token);
 
         const testMicroserviceOne = {
             name: `test-microservice-one`,
@@ -267,7 +271,6 @@ describe('Delete a microservice', () => {
     });
 
     afterEach(async () => {
-        await UserModel.deleteMany({}).exec();
         await MicroserviceModel.deleteMany({}).exec();
         await EndpointModel.deleteMany({}).exec();